Add type hints to add_instance functions
